refactored files to make login and recipe gen work, added additional features to recipe gen
